project/requirements.txt
jupyter==1.0.0
kaggle==1.6.12
numpy==1.26.4
opencv-python==4.9.0.80
matplotlib==3.8.4
scikit-learn==1.4.2
ipywidgets== 8.1.2
pandas==2.2.2
torch==2.2.2
timm-0.9.16